+++ pkg/node_observer/controller.go
@@ -26,6 +26,7 @@ const (
 
 type Controller struct {
 	ctx            context.Context
+	cancel         context.CancelFunc
 	client         kubernetes.Interface
 	statusInformer *StatusInformer
 	healthURL      string
@@ -44,9 +45,10 @@ func NewController(ctx context.Context, client kubernetes.Interface, cfg *Config
 		return nil, err
 	}
 
-	f := httpreq.GetRequestFunc(ctx, http.MethodPost, headers, nil, data, cfg.GenerateTopologyURL)
+	controllerCtx, cancel := context.WithCancel(ctx)
+	f := httpreq.GetRequestFunc(controllerCtx, http.MethodPost, headers, nil, data, cfg.GenerateTopologyURL)
 	statusInformer, err := NewStatusInformer(
-		ctx,
+		controllerCtx,
 		client,
 		&cfg.Trigger,
 		&cfg.APIServer,
@@ -56,10 +58,12 @@ func NewController(ctx context.Context, client kubernetes.Interface, cfg *Config
 		f,
 	)
 	if err != nil {
+		cancel()
 		return nil, err
 	}
 	return &Controller{
-		ctx:            ctx,
+		ctx:            controllerCtx,
+		cancel:         cancel,
 		client:         client,
 		statusInformer: statusInformer,
 		healthURL:      healthURL,
@@ -78,5 +82,6 @@ func (c *Controller) Start() error {
 
 func (c *Controller) Stop(err error) {
 	klog.Infof("Stopping state observer")
+	c.cancel()
 	c.statusInformer.Stop(err)
 }